+++ src/features/set_timeout.rs
@@ -16,36 +16,51 @@ use std::time::Duration;
 /// rt.eval(Script::new("test_timeout.es", "setTimeout(() => {console.log('timed logging')}, 1000);"));
 /// std::thread::sleep(Duration::from_secs(2));
 /// ```
-
 pub fn init(q_js_rt: &QuickJsRuntimeAdapter) -> Result<(), JsError> {
     log::trace!("set_timeout::init");
 
     q_js_rt.add_context_init_hook(|_q_js_rt, q_ctx| {
-        let set_timeout_func =
-            functions::new_native_function_q(q_ctx, "setTimeout", Some(set_timeout), 2, false)?;
-        let set_interval_func =
-            functions::new_native_function_q(q_ctx, "setInterval", Some(set_interval), 2, false)?;
-        let clear_timeout_func =
-            functions::new_native_function_q(q_ctx, "clearTimeout", Some(clear_timeout), 1, false)?;
-        let clear_interval_func = functions::new_native_function_q(
-            q_ctx,
-            "clearInterval",
-            Some(clear_interval),
-            1,
-            false,
-        )?;
-
         let global = unsafe { get_global(q_ctx.context) };
+        #[cfg(feature = "settimeout")]
+        {
+            let set_timeout_func =
+                functions::new_native_function_q(q_ctx, "setTimeout", Some(set_timeout), 2, false)?;
+            let clear_timeout_func = functions::new_native_function_q(
+                q_ctx,
+                "clearTimeout",
+                Some(clear_timeout),
+                1,
+                false,
+            )?;
+            objects::set_property2_q(q_ctx, &global, "setTimeout", &set_timeout_func, 0)?;
+            objects::set_property2_q(q_ctx, &global, "clearTimeout", &clear_timeout_func, 0)?;
+        }
+        #[cfg(feature = "setinterval")]
+        {
+            let set_interval_func = functions::new_native_function_q(
+                q_ctx,
+                "setInterval",
+                Some(set_interval),
+                2,
+                false,
+            )?;
+            let clear_interval_func = functions::new_native_function_q(
+                q_ctx,
+                "clearInterval",
+                Some(clear_interval),
+                1,
+                false,
+            )?;
 
-        objects::set_property2_q(q_ctx, &global, "setTimeout", &set_timeout_func, 0)?;
-        objects::set_property2_q(q_ctx, &global, "setInterval", &set_interval_func, 0)?;
-        objects::set_property2_q(q_ctx, &global, "clearTimeout", &clear_timeout_func, 0)?;
-        objects::set_property2_q(q_ctx, &global, "clearInterval", &clear_interval_func, 0)?;
+            objects::set_property2_q(q_ctx, &global, "setInterval", &set_interval_func, 0)?;
+            objects::set_property2_q(q_ctx, &global, "clearInterval", &clear_interval_func, 0)?;
+        }
         Ok(())
     })?;
     Ok(())
 }
 
+#[cfg(feature = "settimeout")]
 unsafe extern "C" fn set_timeout(
     context: *mut q::JSContext,
     _this_val: q::JSValue,
@@ -107,6 +122,7 @@ unsafe extern "C" fn set_timeout(
     })
 }
 
+#[cfg(feature = "setinterval")]
 unsafe extern "C" fn set_interval(
     context: *mut q::JSContext,
     _this_val: q::JSValue,
@@ -171,6 +187,7 @@ unsafe extern "C" fn set_interval(
     })
 }
 
+#[cfg(feature = "setinterval")]
 unsafe extern "C" fn clear_interval(
     context: *mut q::JSContext,
     _this_val: q::JSValue,
@@ -195,6 +212,7 @@ unsafe extern "C" fn clear_interval(
     })
 }
 
+#[cfg(feature = "settimeout")]
 unsafe extern "C" fn clear_timeout(
     context: *mut q::JSContext,
     _this_val: q::JSValue,