+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * implemented CompiledModuleLoader
 * implemented jsValueAdapter.js_to_str()
 * changed some deps to minor version x.x instead of x.x.x
+* made console/setinterval/settimeout/setimmediate optional(but default) features
 
 # 0.6.0